requirements.txt
boto3==1.34.0
ultralytics==8.0.196
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
python-dotenv==1.0.0
mangum==0.17.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6 